[ADD] マップのコンポーネント化
